+++ src/renderer/index.tsx
@@ -8,6 +8,7 @@ import { KAFKA_CLUSTER_CATALOG_KEY } from "./kafka-cluster-catalog";
 import { KafkaConnectPage, type KafkaConnectPageProps } from "./kafka-connect-pages";
 import { KafkaConnectSettingsStore } from "./kafka-connect-settings";
 import { KafkaConnectionSettingsStore } from "./kafka-connection-settings";
+import { KafkaEndpointSecretsStore } from "./kafka-endpoint-secrets";
 import { KafkaGroupsPage, type KafkaGroupsPageProps } from "./kafka-group-pages";
 import { KafkaIpcRenderer } from "./kafka-ipc";
 import { MANUAL_ENDPOINTS_KEY } from "./kafka-manual-endpoints";
@@ -62,6 +63,7 @@ export default class KafkaExtensionRenderer extends Renderer.LensExtension {
   private readonly writeSettings = new KafkaWriteSettingsStore();
   private readonly schemaRegistrySettings = new KafkaSchemaRegistrySettingsStore();
   private readonly connectSettings = new KafkaConnectSettingsStore();
+  private readonly endpointSecrets = new KafkaEndpointSecretsStore();
   private readonly overviewSettings = new KafkaOverviewSettingsStore();
   private readonly aclAvailability = new KafkaAclAvailabilityStore();
   private readonly hiddenMenu = computed(() => false);
@@ -180,6 +182,7 @@ export default class KafkaExtensionRenderer extends Renderer.LensExtension {
             connectionSettings={this.connectionSettings}
             writeSettings={this.writeSettings}
             schemaRegistrySettings={this.schemaRegistrySettings}
+            endpointSecrets={this.endpointSecrets}
             connectSettings={this.connectSettings}
             kubernetesClusterId={this.clusterId}
             resourceCache={this.resourceCache}
@@ -237,6 +240,7 @@ export default class KafkaExtensionRenderer extends Renderer.LensExtension {
             produce={this.produce}
             deleteTopic={this.deleteTopic}
             schemaRegistrySettings={this.schemaRegistrySettings}
+            endpointSecrets={this.endpointSecrets}
             reachability={this.reachability}
             subscribeProgress={this.subscribeProgress}
           />
@@ -299,6 +303,7 @@ export default class KafkaExtensionRenderer extends Renderer.LensExtension {
             discover={this.discover}
             overview={this.overview}
             schemaRegistrySettings={this.schemaRegistrySettings}
+            endpointSecrets={this.endpointSecrets}
             schemaSubjectNames={this.schemaSubjectNames}
             schemaSubjectDetail={this.schemaSubjectDetail}
             schemaRegister={this.schemaRegister}
@@ -322,6 +327,7 @@ export default class KafkaExtensionRenderer extends Renderer.LensExtension {
             discover={this.discover}
             overview={this.overview}
             connectSettings={this.connectSettings}
+            endpointSecrets={this.endpointSecrets}
             connectNames={this.connectNames}
             connectDetail={this.connectDetail}
             connectPause={(request) => this.client.connectPause(request)}